Fix precision warning.

+++ src/SciTEBase.cxx
@@ -1456,7 +1456,7 @@ void SciTEBase::FillFunctionDefinition(int pos /*= -1*/) {
 		if (words.empty())
 			return;
 		// Counts how many call tips
-		maxCallTips = std::count(words.begin(), words.end(), ' ') + 1;
+		maxCallTips = static_cast<int>(std::count(words.begin(), words.end(), ' ') + 1);
 
 		// Should get current api definition
 		std::string word = apis.GetNearestWord(currentCallTipWord.c_str(), currentCallTipWord.length(),
